requirements.txt
anyio==3.5.0
asgiref==3.5.0
aspose-words==23.8.0
async-generator==1.10
attrs==22.1.0
autopep8==1.7.0
bcrypt==3.2.0
beanie==1.10.4
beautifulsoup4==4.12.2
bokeh==3.0.3
bs4==0.0.1
bson==0.5.10
certifi==2022.9.24
cffi==1.15.0
charset-normalizer==2.1.1
click==8.0.4
colorama==0.4.5
contourpy==1.0.7
cryptography==36.0.2
databricks-cli==0.17.7
decouple==0.0.7
defusedxml==0.7.1
Django==4.1
dnspython==2.2.0
docx2txt==0.8
ecdsa==0.17.0
email-validator==1.1.3
exceptiongroup==1.0.4
fastapi==0.74.1
greenlet==1.1.3
gunicorn==21.2.0
h11==0.13.0
httpie==3.2.1
idna==3.3
Jinja2==3.0.3
lazy-model==0.0.5
lxml==4.9.3
markdown-it-py==2.2.0
MarkupSafe==2.1.0
mdurl==0.1.2
motor==2.5.1
multidict==6.0.2
mysql-connector-python==8.0.31
numpy==1.24.1
oauthlib==3.2.2
outcome==1.2.0
packaging==23.0
pandas==1.5.3
passlib==1.7.4
Pillow==9.4.0
protobuf==3.20.1
pyasn1==0.4.8
pycodestyle==2.9.1
pycparser==2.21
pydantic==1.9.0
Pygments==2.15.0
PyJWT==2.3.0
pymongo==4.5.0
pypandoc==1.11
PySocks==1.7.1
python-dateutil==2.8.2
python-decouple==3.3
python-docx==0.8.11
python-dotenv==0.20.0
python-jose==3.3.0
python-multipart==0.0.5
pytz==2022.7
PyYAML==6.0
requests==2.31.0
requests-toolbelt==0.10.1
rich==13.3.4
rsa==4.8
selenium==4.6.1
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==1.4.32
sqlalchemy2-stubs==0.0.2a20
sqlmodel==0.0.6
sqlparse==0.4.2
starlette==0.17.1
tabulate==0.9.0
toml==0.10.2
tomli==2.0.1
tornado==6.2
trio==0.22.0
trio-websocket==0.9.2
twilio==7.16.0
typing_extensions==4.1.1
tzdata==2022.1
urllib3==1.26.13
uvicorn==0.17.5
wsproto==1.2.0
xyzservices==2022.9.0
yarl==1.7.2

